--- app/src/main/java/org/stepic/droid/util/glide/GlideSvgRequestFactory.java
@@ -1,17 +1,0 @@
-package org.stepic.droid.util.glide;
-
-import android.content.Context;
-import android.graphics.drawable.Drawable;
-import android.graphics.drawable.PictureDrawable;
-
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.RequestBuilder;
-
-public class GlideSvgRequestFactory {
-    public static RequestBuilder<PictureDrawable> create(Context context, Drawable placeholder) {
-        return Glide.with(context.getApplicationContext())
-            .as(PictureDrawable.class)
-            .placeholder(placeholder)
-            .listener(new SvgSoftwareLayerSetter());
-    }
-}
