+++ app/src/main/java/org/stepik/android/view/glide/mapper/SvgSoftwareLayerSetter.java
@@ -1,4 +1,4 @@
-package org.stepic.droid.util.glide;
+package org.stepik.android.view.glide.mapper;
 
 
 import android.annotation.TargetApi;
